refactor: separated interfaces from presentation in two files, controller and router
