+++ src/presentation/controller/AbstractController.ts
@@ -1,5 +1,5 @@
 import { IEndPointsService } from 'interfaces/domain'
-import { IEndPointsController } from 'interfaces/presentation'
+import { IEndPointsController } from 'interfaces/presentation/controller'
 
 export default abstract class AbstractController implements IEndPointsController {
   service: IEndPointsService
